Prevent re-sending source strings by default.
This change ensures that for normal submission, source documents that have
already been sent to the server will not be sent again. This is achieved
by adding a remote mapping to the job item upon successful submission.
Presence of a remote mapping item is taken to imply prior submission.

An extra button is added to force resubmit for all items regardless of
prior submission.