Segment content into simple paragraphs.
This splits text on the sequence "\r\n\r\n" as an approximation
of paragraphs before sending for translation. Segments that would
be empty or made of only whitespace are not sent for translation.
All content around translatable segments is preserved.

WIP: segmentation. Squash further changes into this.

Move segmentation code to a separate class.