Prevent actions on jobs that are not currently active.
The behaviour is undefined for fetching translations after a job has
been accepted, so the actions are disabled after the job is accepted
so that users are not confused by undefined behaviour. In particular,
translations could be updated in the translation job, but would never
be copied to the published translation.